move focal loss to classification.
