+++ optax/losses/classification_test.py
@@ -469,5 +469,22 @@ class CTCTest(parameterized.TestCase):
           jnp.array(expected_loss), per_seq_loss[n], rtol=self._rtol)
 
 
+class SigmoidFocalLossTest(parameterized.TestCase):
+
+  def setUp(self):
+    super().setUp()
+    self.ys = np.array([[2., 1., -2.], [.3, 2., 1.2]], dtype=np.float32)
+    self.ts = np.array([[0., 0., 1.], [1., 0., 0.]])
+
+  @chex.all_variants
+  def test_gamma_zero(self):
+    """From gamma == 0 we expect a CE loss."""
+    np.testing.assert_allclose(
+      self.variant(
+        classification.sigmoid_focal_loss)(self.ys, self.ts, gamma=0.),
+        classification.sigmoid_binary_cross_entropy(self.ys, self.ts),
+        atol=1e-4)
+
+
 if __name__ == '__main__':
   absltest.main()
